All done!
Opted to use an advanced chart instead of 2 separate charts and the radar chart. With help from Tina, I've targeted the datasets **as an array**, giving me the desired animation. As such, I removed some unnecessary HTML and CSS that was used on the radar chart. This concludes my Data Viz assignment!